requirements.txt
streamlit
numpy
Pillow
scikit-learn
opencv-python-headless
torch==2.0.1
torchvision==0.15.2
